+++ lib/extended_math.dart
@@ -13,9 +13,12 @@ export 'src/exceptions/matrix_exception.dart';
 export 'src/exceptions/vector_exception.dart';
 
 export 'src/matrix/base/matrix_base.dart';
+export 'src/matrix/diagonal_matrix.dart';
 export 'src/matrix/matrix.dart';
 export 'src/matrix/square_matrix.dart';
+
 export 'src/tensor/base/tensor_base.dart';
 export 'src/tensor/tensor.dart';
+
 export 'src/vector/base/vector_base.dart';
 export 'src/vector/vector.dart';