+++ test/extended_math_test.dart
@@ -1,15 +1,10 @@
+import 'dart:math';
+
 import 'package:extended_math/extended_math.dart';
 
 void main() {
+  final v1 = Vector(<double>[3, 1, 1]);
+  final v2 = Vector(<double>[-2, 5, 1]);
 
-  final m = Matrix(<List<double>>[
-    <double>[2, -1, 5],
-    <double>[0, 2, 1],
-    <double>[3, 1, 1]
-  ]);
-  print(m.frobeniusNorm());
-
-  final v1 = Vector(<double>[6, 3]);
-  final v2 = Vector(<double>[5, 13]);
-  print(v1.angleBetween(v2));
+  print(v1.dotProduct(v2));
 }